# apps/backend/docs/public-api/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Scheduler Public API
  version: 1.0.0
  description: |
    External-facing REST API for third-party integrations with the Social Media Scheduler platform.
    
    ## Authentication
    
    All requests require an API key passed in the `X-API-Key` header:
    
    ```
    X-API-Key: sk_live_your_api_key_here
    ```
    
    API keys can be created and managed through the dashboard at `/settings/api-keys`.
    
    ## Rate Limiting
    
    API requests are rate-limited per API key. Default limits:
    - 1000 requests per hour per API key
    - 5000 requests per hour per workspace (across all keys)
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed in the window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    
    When rate limited, you'll receive a `429 Too Many Requests` response with a `Retry-After` header.
    
    ## Scopes & Permissions
    
    API keys have granular permission scopes that control access to endpoints:
    
    - `posts:read` - Read posts and drafts
    - `posts:write` - Create, update, and delete posts
    - `analytics:read` - Read analytics data
    - `media:read` - Read media library
    - `media:write` - Upload and delete media
    - `accounts:read` - Read connected social accounts
    - `accounts:write` - Connect and disconnect accounts
...
    
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/public/v1/media:
    get:
      tags:
        - Media
      summary: List media files
      description: Retrieve a paginated list of media files for your workspace
      operationId: listMedia
      security:
        - ApiKeyAuth: [media:read]
      parameters:
        - name: type
          in: query
          description: Filter by media type
          required: false
          schema:
            type: string
            enum: [image, video]
        - name: limit
          in: query
          description: Number of items to return (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  media:
                    type: array
                    items:
                      $ref: '#/components/schemas/Media'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    
    post:
      tags:
        - Media
      summary: Upload media file
      description: Upload a new media file (Not yet implemented)
      operationId: uploadMedia
      security:
        - ApiKeyAuth: [media:write]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '501':
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/public/v1/media/{id}:
    get:
      tags:
        - Media
      summary: Get a media file
      description: Retrieve a single media file by ID
      operationId: getMedia
      security:
        - ApiKeyAuth: [media:read]
      parameters:
        - name: id
          in: path
          description: Media ID
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{24}$'
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  media:
                    $ref: '#/components/schemas/Media'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    
    delete:
      tags:
        - Media
      summary: Delete a media file
      description: Delete a media file by ID
      operationId: deleteMedia
      security:
        - ApiKeyAuth: [media:write]
      parameters:
        - name: id
          in: path
          description: Media ID
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{24}$'
      responses:
        '200':
          description: Media deleted successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Media deleted successfully
                  mediaId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Format: `sk_live_<random_string>`
        
        Create API keys in your dashboard at `/settings/api-keys`.
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        example: 999
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
        example: 1709812800
  schemas:
...
    Media:
      type: object
      properties:
        id:
          type: string
          description: Unique media identifier
        filename:
          type: string
          description: Original filename
        mediaType:
          type: string
          enum: [image, video]
          description: Type of media file
        storageUrl:
          type: string
          description: Storage URL (internal)
        cdnUrl:
          type: string
          description: CDN URL for public access
        thumbnailUrl:
          type: string
          nullable: true
          description: Thumbnail URL (for videos)
        size:
          type: integer
          description: File size in bytes
        mimeType:
          type: string
          description: MIME type
          example: image/jpeg
        width:
          type: integer
          nullable: true
          description: Image/video width in pixels
        height:
          type: integer
          nullable: true
          description: Image/video height in pixels
        duration:
          type: number
          nullable: true
          description: Video duration in seconds
        status:
          type: string
          enum: [uploading, processing, ready, failed]
          description: Processing status
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        pages:
          type: integer
          description: Total number of pages
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: Unauthorized
        message:
          type: string
          description: Human-readable error message
          example: Invalid API key
        code:
          type: string
          description: Machine-readable error code
          example: INVALID_API_KEY
        requestId:
          type: string
          description: Unique request identifier for debugging
          example: req_abc123xyz
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        docsUrl:
          type: string
          description: Link to error documentation
          example: https://docs.example.com/errors/INVALID_API_KEY
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            message: Content, socialAccountIds, and scheduledAt are required
            code: VALIDATION_ERROR
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid API key
            code: INVALID_API_KEY
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/INVALID_API_KEY
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: 'Missing required scopes: posts:write'
            code: INSUFFICIENT_SCOPE
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/INSUFFICIENT_SCOPE
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Post not found
            code: NOT_FOUND
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
            example: 3600
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Too Many Requests
            message: Rate limit exceeded. Please retry after 3600 seconds.
            code: RATE_LIMIT_EXCEEDED
            requestId: req_abc123xyz
            timestamp: '2026-03-07T10:30:00Z'
            docsUrl: https://docs.example.com/errors/RATE_LIMIT_EXCEEDED